+++ crates/config/src/template_config.rs
@@ -31,6 +31,7 @@ macro_rules! var_setting {
     };
 }
 
+var_setting!(TemplateVariableArraySetting, Vec<Value>);
 var_setting!(TemplateVariableBoolSetting, bool);
 var_setting!(TemplateVariableNumberSetting, isize);
 var_setting!(TemplateVariableObjectSetting, FxHashMap<String, Value>);
@@ -175,6 +176,10 @@ config_enum!(
     #[derive(Config)]
     #[serde(tag = "type", expecting = "expected a supported value type")]
     pub enum TemplateVariable {
+        /// An array variable.
+        #[setting(nested)]
+        Array(TemplateVariableArraySetting),
+
         /// A boolean variable.
         #[setting(nested)]
         Boolean(TemplateVariableBoolSetting),
@@ -200,6 +205,7 @@ config_enum!(
 impl TemplateVariable {
     pub fn get_order(&self) -> usize {
         let order = match self {
+            Self::Array(cfg) => cfg.order.as_ref(),
             Self::Boolean(cfg) => cfg.order.as_ref(),
             Self::Enum(cfg) => cfg.order.as_ref(),
             Self::Number(cfg) => cfg.order.as_ref(),
@@ -212,6 +218,7 @@ impl TemplateVariable {
 
     pub fn is_internal(&self) -> bool {
         match self {
+            Self::Array(cfg) => cfg.internal,
             Self::Boolean(cfg) => cfg.internal,
             Self::Enum(cfg) => cfg.internal,
             Self::Number(cfg) => cfg.internal,
@@ -229,6 +236,7 @@ impl TemplateVariable {
 
     pub fn is_required(&self) -> bool {
         match self {
+            Self::Array(cfg) => cfg.required,
             Self::Boolean(cfg) => cfg.required,
             Self::Number(cfg) => cfg.required,
             Self::Object(cfg) => cfg.required,