+++ packages/types/src/template-config.ts
@@ -21,6 +21,21 @@ export interface PartialTemplateFrontmatterConfig {
 }
 
 /** Configuration for a template variable. */
+export interface TemplateVariableArraySetting {
+	/** The default value of the variable if none was provided. */
+	default: unknown[];
+	/** Marks the variable as internal, and won't be overwritten via CLI arguments. */
+	internal: boolean;
+	/** The order in which variables should be prompted for. */
+	order: number | null;
+	/** Prompt the user for a value when the generator is running. */
+	prompt: string | null;
+	/** Marks the variable as required, and will not accept an empty value. */
+	required: boolean | null;
+	type: 'array';
+}
+
+/** Configuration for a template variable. */
 export interface TemplateVariableBoolSetting {
 	/** The default value of the variable if none was provided. */
 	default: boolean;
@@ -108,6 +123,7 @@ export interface TemplateVariableStringSetting {
 }
 
 export type TemplateVariable =
+	| TemplateVariableArraySetting
 	| TemplateVariableBoolSetting
 	| TemplateVariableEnumSetting
 	| TemplateVariableNumberSetting
@@ -139,6 +155,21 @@ export interface TemplateConfig {
 	 * Variables can also be populated by passing command line arguments.
 	 */
 	variables: Record<string, TemplateVariable>;
+}
+
+/** Configuration for a template variable. */
+export interface PartialTemplateVariableArraySetting {
+	/** The default value of the variable if none was provided. */
+	default?: unknown[] | null;
+	/** Marks the variable as internal, and won't be overwritten via CLI arguments. */
+	internal?: boolean | null;
+	/** The order in which variables should be prompted for. */
+	order?: number | null;
+	/** Prompt the user for a value when the generator is running. */
+	prompt?: string | null;
+	/** Marks the variable as required, and will not accept an empty value. */
+	required?: boolean | null;
+	type?: 'array' | null;
 }
 
 /** Configuration for a template variable. */
@@ -229,6 +260,7 @@ export interface PartialTemplateVariableStringSetting {
 }
 
 export type PartialTemplateVariable =
+	| PartialTemplateVariableArraySetting
 	| PartialTemplateVariableBoolSetting
 	| PartialTemplateVariableEnumSetting
 	| PartialTemplateVariableNumberSetting